requirements.txt
configparser
openai
requests
python-docx
python-dotenv